+++ libawui/awui/IO/Stream.h
@@ -9,10 +9,10 @@ namespace awui {
 			public:
 				virtual void Close() = 0;
 
-				virtual void SetPosition(int64_t value) = 0;
-				virtual int64_t GetPosition() = 0;
+				virtual void SetPosition(uint32_t value) = 0;
+				virtual uint32_t GetPosition() = 0;
 
-				virtual int64_t GetLength() = 0;
+				virtual uint32_t GetLength() = 0;
 
 				virtual uint8_t ReadByte() = 0;
 				virtual void WriteByte(uint8_t value) = 0;
